% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse_saekernel.R
\name{mse_saekernel}
\alias{mse_saekernel}
\title{Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel and Bootstrap Mean Squared Error Estimators}
\usage{
mse_saekernel(X, Y, vardir, bandwidth, B = 1000)
}
\arguments{
\item{X}{Auxiliary variable of X}

\item{Y}{Direct Estimation of Y}

\item{vardir}{Sampling variances of direct estimators}

\item{bandwidth}{The kernel bandwidth smoothing parameter}

\item{B}{Number of bootstrap. Default is 1000}
}
\value{
This function returns a list with following objects:
\item{est}{a value of Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel estimators}
\item{refvar}{estimated random effect variance}
\item{mse}{estimated Bootstrap Mean Squared Error Estimators of Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel}
}
\description{
This Function Gives Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel and Calculates the bootstrap mean squared error estimates
}
