% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saekernel.R
\name{saekernel}
\alias{saekernel}
\title{Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel}
\usage{
saekernel(X, Y, vardir, bandwidth, K = dnorm)
}
\arguments{
\item{X}{Auxiliary variable of X}

\item{Y}{Direct Estimation of Y}

\item{vardir}{Sampling variances of direct estimators}

\item{bandwidth}{The kernel bandwidth smoothing parameter}

\item{K}{The Kernel to be used. Only can be "dnorm" or "dunif"}
}
\value{

}
\description{
This Function Gives Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel
}
