% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saekernel.R
\name{saekernel}
\alias{saekernel}
\title{Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel}
\usage{
saekernel(X, Y, vardir, bandwidth)
}
\arguments{
\item{X}{Auxiliary variable of X}

\item{Y}{Direct Estimation of Y}

\item{vardir}{Sampling variances of direct estimators}

\item{bandwidth}{The kernel bandwidth smoothing parameter}
}
\value{
This function returns a list with following objects:
\item{est}{a value of Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel estimators}
\item{refvar}{estimated random effect variance}
}
\description{
This Function Gives Small Area Estimation Non-Parametric based Nadaraya-Watson Kernel
}
